Move plugin loading to dispatcher startup
Move credential type and inventory plugin loading out of Django's
app.ready() and into the dispatcher's on-start producer task. This
avoids database operations during management commands that don't
need them (like collectstatic) and ensures plugins are loaded after
the dispatcher service is fully initialized.

- Remove _load_credential_types_feature, load_credential_types_feature,
  and load_inventory_plugins from MainConfig.ready()
- Add load_inventory_plugins() and load_credential_types_feature() to
  awx.main.tasks.system, called during dispatch_startup
- Use advisory locks to prevent duplicate loading across workers
- Move update_inventory_computed_fields and handle_removed_image imports
  in signals.py to be lazy to avoid circular imports
- Remove bypass_in_test decorator (no longer needed)
